dash: parse version-1 sidx (64-bit) + fail loud on zero-segment tracks
A stream delivered via SegmentBase + indexRange played nothing: the player
fetched manifest, init segments and keys but never requested a single media
segment (interceptor saw kind=2 = 0) — eternal buffering, no error.

Root cause: parse_sidx read earliest_presentation_time + first_offset as
32-bit unconditionally, ignoring the sidx `version`. In a version-1 sidx
those fields are 64-bit (ISO/IEC 14496-12 §8.16.3), so every later field
shifted by 8 bytes and entry_count came out garbage (typically 0) → empty
segment list → no media scheduled, silently. A version-0 sidx (the smoke
stream) was unaffected, which is why that one played.

- parse_sidx: read version, then 64-bit EPT/first_offset for version >= 1
  (32-bit otherwise); add read_u64 + a `size == 1` largesize guard; widen
  SidxBox.{earliest_presentation_time,first_offset} and Segment::new's time
  bases to u64. Log `[sidx] version/entries/...` so the failure mode is
  visible in one line.
- play(): fail loud (PlayerEvent::Error, ManifestParse) when the selected
  video/audio representation has zero media segments, instead of leaving the
  pipeline waiting forever for frames that never arrive.

+++ player/src/tracks/segment.rs
@@ -26,8 +26,8 @@ impl Segment {
         file_url: &String,
         start: u64,
         end: u64,
-        start_time_base: Option<u32>,
-        end_time_base: Option<u32>,
+        start_time_base: Option<u64>,
+        end_time_base: Option<u64>,
         timescale: Option<u32>,
     ) -> Result<Self, Box<dyn Error>> {
         let timescale = match timescale {
@@ -35,11 +35,11 @@ impl Segment {
             None => 0,
         };
 
-        let to_duration = |time: u32| -> Duration {
+        let to_duration = |time: u64| -> Duration {
             if timescale == 0 {
                 Duration::ZERO
             } else {
-                Duration::from_micros((time as u64) * 1_000_000 / (timescale as u64))
+                Duration::from_micros(time * 1_000_000 / (timescale as u64))
             }
         };
